requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.1
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.6
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
